+++ frb_codegen/src/library/commands/format_dart.rs
@@ -3,17 +3,24 @@ use crate::commands::command_runner::call_shell;
 use crate::library::commands::command_runner::check_exit_code;
 use crate::utils::path_utils::{normalize_windows_unc_path, path_to_string};
 use anyhow::Context;
+use anyhow::Result;
+use itertools::Itertools;
 use log::debug;
 use pathdiff::diff_paths;
 use std::path::{Path, PathBuf};
 
 #[allow(clippy::vec_init_then_push)]
-pub fn format_dart(paths: &[PathBuf], base_path: &Path, line_length: u32) -> anyhow::Result<()> {
+pub fn format_dart(
+    paths: &[PathBuf],
+    base_path: &Path,
+    line_length: u32,
+    extra_extensions: &[&str],
+) -> anyhow::Result<()> {
     if paths.is_empty() {
         return Ok(());
     }
 
-    let paths = prepare_paths(paths, base_path)?;
+    let paths = prepare_paths(paths, base_path, extra_extensions)?;
     debug!("execute format_dart paths={paths:?} line_length={line_length}");
 
     let res = command_run!(
@@ -28,11 +35,15 @@ pub fn format_dart(paths: &[PathBuf], base_path: &Path, line_length: u32) -> any
     Ok(())
 }
 
-pub(super) fn prepare_paths(paths: &[PathBuf], base_path: &Path) -> anyhow::Result<Vec<PathBuf>> {
+pub(super) fn prepare_paths(
+    paths: &[PathBuf],
+    base_path: &Path,
+    extra_extensions: &[&str],
+) -> anyhow::Result<Vec<PathBuf>> {
     let base_path_str = path_to_string(base_path)?;
     let normalized_base_path = normalize_windows_unc_path(&base_path_str);
 
-    (paths.iter())
+    Ok((paths.iter())
         .map(|path| {
             let mut path: PathBuf =
                 (normalize_windows_unc_path(&path_to_string(path)?).to_owned()).into();
@@ -42,5 +53,16 @@ pub(super) fn prepare_paths(paths: &[PathBuf], base_path: &Path) -> anyhow::Resu
             }
             Ok(path)
         })
-        .collect()
+        .collect::<Result<Vec<_>>>()?
+        .into_iter()
+        .flat_map(|path| {
+            (candidate_extensions.iter()).map(move |ext| with_extension(path.clone(), ext))
+        })
+        .filter(|path| path.exists())
+        .collect_vec())
+}
+
+fn with_extension(mut path: PathBuf, ext: &str) -> PathBuf {
+    path.set_extension(ext);
+    path
 }
